requirements.txt
discord.py==2.4.0
PyAutoGUI==0.9.54
selenium==4.24.0
pyyaml==6.0.2
